added GaP to the materials libary
changed all elastic constants to N/m^2 to ensure the AlloyAB class works correctly with the existing materials. 

NOTE: CdTe PbTe PbSe need to be changed by Eugen! Do not know where the values were taken from and what are the current units.

added a RelaxationTrianlge method to the AlloyAB class. It is implemented as general as possible but works only for cubic materials. Should work for arbitrary surface orientations, was tested for (001) and (111) only. Anyhow double check the results!