NOTE: changed the default value of the trans flag of HXRD.Q2Ang method to true because this is necessary for coplanar scattering experiments to get the correct angles
included support for high and low incidence geometry in the Q2Ang method of HXRD
the delta angle returned by HXRD.Q2Ang may not be correct by 180 degree, this needs to be checked.

added some input data tests to the experiment class to ensure that useful inplane and out-of plane directions are given by the user and warn the user in case of controversies.