This commit includes major changes to the xrutils.experiment classes, which includes slight interface changes. The conversion from angular to reciprocal space is now implemented in a separate class, which can be initialized with the goniometer setup to provide the suiteable routines for conversion of angular coordinates to reciprocal space. Special routines for linear and area detectors exist, which must be initialized with detector geometry(either distance of ch/deg and mounting direction). The working routines are implemented in C-code to be performant. In addition to this changes the documentation was updated and includes a special section on how the conversion routines work.
new functionality was added in the normalize.py file:
 this includes:
 * blockaverage functions for 1D,2D and PSD arrays, including the
   possibility to apply regions of interest.
 * IntensityNormalizer class for the absorber/count time/monitor correction
   of intensities

The block average routines and qconversion are coded in C. C code files are therefore added to the repository.

Illustions on how to use xrutils, and one example was added in the documentation.
The SConstruct build file was changed to check on dependencies prior to building and a debug command line argument was added to scons (scons debug=1: uses debug flags for building now)