utilities.py file was added with a
* lam2en function converting an energy to a wavelength
* and a maplog function useful for plotting

the lam2en function is now used in the Experiment and Material classes.
the TODO file was updated.

A acurate value for the Copper Ka1 wavelength is now used in the Experiment class
I do not know where Eugen found the value he put there! Now a very acurate value 
from the International Tables of Crystallography is used (exact reference in the code).