fix memory leak in c-code
several c-functions had a problem with reference counting. This was introduced
recently with a change enable that instead of numpy arrays also list can be
given to the c-code. the function PyArray_FROM_OTF increased the reference
counter which has to be manually decreased in the end of a function.

Usage of memory_profiler helped to find this issue!