add FuzzyGridder2D class and its unittest
the concept of fuzzy gridding which allows assigning a certain data width to
each data point was extended to the 2D case. Although this adds some execution
time to the 2D gridder this might be a small price to pay in order to get less
dependent on the number of grid points and might enable some edge case
applications to work with gridded data.