improve performance of XRD simulations
simulation speed was significantly increased. this was achieved by
optimizing the calculation of the chi0/chih/structure factors and by
vectorizing the dynamical model which now does not include a loop over
the q positions

In total the kinematical code now is approximately by a factor of 2.5
and the DynamicalModel by a factor of **25** faster.