fix issue #34, possible fix for issue #35
convolvers are now created as needed and reused as long as possible.
the FP_profile got a class attribute 'isotropic' which specifies if all
the convolvers in the class are isotropic or not. If this is the case
different peaks at the same diffraction angle are combined in
PowderDiffraction, otherwise every peak is treated independently.