rewrite line cut functions (fixes #1)
the new variants work for 2D and 3D data. it can operate on the
ungridded raw data and therefore obtain much better resolution.
furthermore it is faster and the implementation is only a few lines and
should therefore be much easier to debug!

The interface is however incompatible with the old variants. Change of
user-code might be needed and has to be highlighted in the next release