allow '?' in cif file as element name
the 'dummy' element is now used when a '?' is encountered. This is
mostly to prevent the parser from raising an error and often happens
during parsing incomplete data sets in cif files. since often at least
one complete data set is present its preferential that the parser does
not stop upon encountering '?'