better doc/options for PowderDiffraction cleanup
The multiprocessing module did not cleanup upon change of context. This
is somehow the standard behavior and caused problems as described in #61

There is no real fix for this, but the default is now not to initialize
the multiprocessing code in PowderDiffraction (an option was added to
allow for that). And explicit close() methods where added to
PowderDiffraction and PowderModel which should be used at the
end-of-life of an instance.