loosen numerical requirement on fit in unittest
commit 394d36f already reduced the requirement but clearly not enough.
Now I require the test to achieve an error smaller than the data point
spacing, which is still a quite stringend requirement, but before it was
clearly too much and produced several false positive failing unit tests.

It might still fail now. If the failure looks like a numerical problem
it could still be that we assume too high accuracy for the numerical
fit.