enable Panalytical XRDML version >2.0
parsing of new XRDML files needs to read the "counts" field and correct
it by the "beamAttenuationfactor" instead of the old "intensity" field.

The change is a bit a hack and it might be better to introduce a version
detection before parsing the file and make parsing decisions based on
that.