refactor: update concentration checks to use assertions for better error handling
It is also important to assert before setting the value. In this case not a lot of bad things happen if you wrongly set the array, but even if the AssertionError is raised, you did actually put the wrong concentrations into the object and nobody would stop you from continuing down the wrong path.