update readme with respect to non-default install options
disabling openmp support is still possible, but requires a different
command now. Also in case it is enabled an installation will now fail.
This will affect only users which do not use prebuilt wheel packages,
which should be a minority and hopefully typically affect people which
know how to deal with this problem.