fix(examples): set nthreads in example script (#217)
* fix(examples): set nthreads in example script

It seems tests on windows can stall if the number of threads is automatically determined. This might be a limit on the test systems and in my virtual machine, but generally seems an important thing to make the user aware of!

* fix blockaverage2d c-code

recently a bug was introduced which only showed up in multi threaded applications.

* fix linting error due to too long line

One might really discuss if the currently (by the linter) enforced line
length is still good.